chore: move to application components, more sizes of avatar and use the current user as default avatar
