--- client/src/routes/(app)/components/Navbar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Navbar.svelte';
